admin: logic for removing not granted items from menu is now moved to the framework package
